Visualization of data along with the creation of backend features to add Saving amounts

+++ Backend/models/AddExpense.js
@@ -16,7 +16,7 @@ module.exports = (sequelize, DataTypes) => {
             allowNull: false,
         },
         price: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.FLOAT(8,2),
             allowNull: false,
         },
         email: {
